requirements.txt
anthropic
jsonschema>=4.26.0
openpyxl>=3.1.5
pandas>=3.0.1
pytest>=9.0.2
PyYAML>=6.0.3
